requirements.txt
click>=8.1.0
boto3>=1.28.0
pyyaml>=6.0.0
python-terraform>=0.10.1